feat: novo retrieve view e winner packager
