+++ src/services/api/serializers.py
@@ -76,17 +76,6 @@ class CreatePresenceSerializer(serializers.ModelSerializer):
         presence = Presence.objects.create(student=student, talk=talk, **validated_data)
         return presence
 
-class DrawWinnerSerializer(serializers.ModelSerializer):
-    student = serializers.PrimaryKeyRelatedField(queryset=Student.objects.all())
-    talk = serializers.PrimaryKeyRelatedField(read_only=True)
-
-    class Meta:
-        model = DrawWinner
-        fields = ['id', 'talk', 'student']
-
-    def create(self, validated_data):
-        draw_winner = DrawWinner.objects.create(**validated_data)
-        return draw_winner
 
 
 class AdminSerializer(serializers.Serializer):
