+++ tests/test_adminListCreateWinner.py
@@ -3,7 +3,8 @@ from rest_framework.test import APITestCase
 from rest_framework import status
 from django.contrib.auth.models import User
 from rest_framework.test import APIClient
-from services.api.models import Talk, Presence, DrawWinner
+from services.api.models import Talk, Presence
+from services.winners.models import DrawWinner
 from services.students.models import Student
 from services.speakers.models import Speaker
 from datetime import datetime as dt, timedelta
