+++ tests/test_adminListWinnerView.py
@@ -6,7 +6,8 @@ from django.urls import reverse
 from django.test import TestCase
 from datetime import datetime as dt, timedelta
 from zoneinfo import ZoneInfo
-from services.api.models import Talk, DrawWinner
+from services.api.models import Talk
+from services.winners.models import DrawWinner
 from services.students.models import Student
 from services.speakers.models import Speaker
