+++ tests/test_adminRetrieveWinnerByTalkView.py
@@ -3,7 +3,7 @@ from rest_framework.test import APITestCase
 from rest_framework import status
 from django.contrib.auth.models import User
 from rest_framework.test import APIClient
-from services.api.models import Talk, Presence
+from services.api.models import Talk
 from services.winners.models import DrawWinner
 from services.students.models import Student
 from services.speakers.models import Speaker
@@ -13,7 +13,7 @@ import uuid
 
 DATETIME_FORMAT = "%Y-%m-%dT%H:%M"
 
-class AdminRetrieveWinnerByStudentViewTestCase(APITestCase):
+class AdminRetrieveWinnerByTalkViewTestCase(APITestCase):
     def setUp(self):
         self.client = APIClient()
         User.objects.create_superuser(
@@ -23,17 +23,7 @@ class AdminRetrieveWinnerByStudentViewTestCase(APITestCase):
         )
         self.client.login(username='admin', password='password123')
 
-        self.student = Student.objects.create(
-            name="CO-SSI da Silva Junior",
-            email="co-ssi.jr@usp.br",
-            usp_number="281905",
-            code="A001"
-        )
-
-        self.url = reverse('admin-retrieve-winner-by-student', kwargs={'student_id': self.student.id})
-
-    def test_retrieve_draw_winners_by_student(self):
-        speaker = Speaker.objects.create(
+        self.speaker = Speaker.objects.create(
             name="Maria Silva",
             description="Especialista em IA",
             social_media="@maria",
@@ -42,40 +32,39 @@ class AdminRetrieveWinnerByStudentViewTestCase(APITestCase):
         )
 
         datetime_now = dt.now(ZoneInfo('America/Sao_Paulo'))
-        talk1 = Talk.objects.create(
+        self.talk = Talk.objects.create(
             title='Introdução a Machine Learning',
-            speaker=speaker,
+            speaker=self.speaker,
             description='Aprenda o que é Machine Learning e quais técnicas aplicar em cada caso',
             start_time=datetime_now.strftime(DATETIME_FORMAT),
             end_time=(datetime_now + timedelta(hours=2)).strftime(DATETIME_FORMAT),
         )
 
-        talk2 = Talk.objects.create(
-            title='Técnica RandomForest',
-            speaker=speaker,
-            description='Como aplicar a técnica de RandomForest ao seu projeto de IA',
-            start_time=(datetime_now + timedelta(days=1)).strftime(DATETIME_FORMAT),
-            end_time=(datetime_now + timedelta(days=1, hours=2)).strftime(DATETIME_FORMAT),
+        self.url = reverse('admin-retrieve-winner-by-talk', kwargs={'talk_id': self.talk.id})
+
+    def test_retrieve_draw_winners_by_talk(self):
+        student1 = Student.objects.create(
+            name="CO-SSI da Silva Junior",
+            email="co-ssi.jr@usp.br",
+            usp_number="281905",
+            code="A001"
         )
 
-        _presence1 = Presence.objects.create(
-            student=self.student,
-            talk=talk1,
-        )
-
-        _presence2 = Presence.objects.create(
-            student=self.student,
-            talk=talk2,
+        student2 = Student.objects.create(
+            name="Ana Santos",
+            email="ana.santos@usp.br",
+            usp_number="281906",
+            code="A002"
         )
 
         draw_winner1 = DrawWinner.objects.create(
-            talk=talk1,
-            student=self.student,
+            talk=self.talk,
+            student=student1,
         )
 
         draw_winner2 = DrawWinner.objects.create(
-            talk=talk2,
-            student=self.student,
+            talk=self.talk,
+            student=student2,
         )
 
         response = self.client.get(self.url)
@@ -83,31 +72,31 @@ class AdminRetrieveWinnerByStudentViewTestCase(APITestCase):
         self.assertEqual(response.status_code, status.HTTP_200_OK)
         self.assertEqual(len(response.data), 2)
 
-        draw_winner1 = {
+        expected_winner1 = {
             "id": draw_winner1.id,
-            "student": self.student.id,
-            "talk": talk1.id,
+            "student": student1.id,
+            "talk": self.talk.id,
         }
 
-        draw_winner2 = {
+        expected_winner2 = {
             "id": draw_winner2.id,
-            "student": self.student.id,
-            "talk": talk2.id,
+            "student": student2.id,
+            "talk": self.talk.id,
         }
 
-        self.assertDictEqual(draw_winner1, response.data[0])
-        self.assertDictEqual(draw_winner2, response.data[1])
+        self.assertDictEqual(expected_winner1, response.data[0])
+        self.assertDictEqual(expected_winner2, response.data[1])
 
-    def test_retrieve_draw_winners_by_student_invalid_id(self):
-        fake_id = uuid.uuid4()
-        url_with_invalid_id = reverse('admin-retrieve-winner-by-student', kwargs={'student_id': fake_id})
+    def test_retrieve_draw_winners_by_talk_not_found(self):
+        fake_id = 404
+        url_with_invalid_id = reverse('admin-retrieve-winner-by-talk', kwargs={'talk_id': fake_id})
 
         response = self.client.get(url_with_invalid_id)
 
-        self.assertEqual(response.status_code, status.HTTP_400_BAD_REQUEST)
+        self.assertEqual(response.status_code, status.HTTP_404_NOT_FOUND)
         self.assertIn("error", response.data)
 
-    def test_retrieve_draw_winners_by_student_unauthorized(self):
+    def test_retrieve_draw_winners_by_talk_unauthorized(self):
         self.client.logout()
         response = self.client.get(self.url)
         self.assertEqual(response.status_code, status.HTTP_403_FORBIDDEN)
